Completely reset database migration history
This kindof sucks because we'll lose a lot of migration history, but the
reason for doing it is that the schema definitions seems to be out of
sync with the production database.

Not sure how this happened, we probably did something wrong in
development.

If someone ever sees a reason to revert this reset, simply merge the
current migration state (the one before this commit) with the future
one.