Remove caching that wouldn't work anyway
You can't store things in global state in this type of webapp. If you
want to retain a list of keys, you'll have to save it in some type of
state - the DB, on the filesystem, or something like that.

This code hits the local DB which isn't *that* big of a penalty, since
the query is only run once per request, so for now, we'll live without
caching, but might reconsider this at a later point.