Don't filter actor members on stop_date
Another Focus quirk. It looks like some members have stop_date set.
We're not 100% sure of the semantics of this, but it probably means that
their membership isn't paid. We still want to find and see those users.

However, in commit a711cbf62 I filtered the Actor.membership_type()
method on stop_date. This was as far as I can remember because some
members have multiple membership services, e.g. when they age, but
instead of deleting the old one, stop_date is set.