Disable personal membership checks temporarily
We're dependent on the underlying back-end datatypes in Focus in order
to perform lookups with DISTINCT. We've requested a fix, but until that
happens, we'll have to disable these checks.

Pros is that members with multiple membership services (where one is
stopped) will not get errors, cons is that entries that are NOT personal
memberships will not be excluded from Sherpa's functionality.