Use data-dnt-row attribute only for content rows
Previously, data-dnt-row was set on add-content controls that were
contained in a separate row (not included in a row with actual content),
and the attribute was used to separate them.

Don't do that; use the fact that the parent row will already have a
data-dnt-row attribute if it contains content.