Use insecure endpoint for Turbasen
We're experiencing instability with the secure endpoint and therefore
switching to the insecure one for now. The API key will be transmitted
in cleartext, but an attack will have to be specifically targeted and is
considered unlikely.

Note that we should generate a new API key when the secure endpoint
issues have been resolved.