Return user from memberid verification, not actor
A lot of the logic performed on lookup is actually already in place in
the lower-level User.get_or_create_inactive and User.create_inactive
static methods. For example, an expired user who has reappeared in
either the Actor or Enrollment tables will have their expired state
reset to False.

Returning an inactive user allows the caller to see if it's a pending
membership or not.