Fix existing parent references
Pages on new sites actually have parent references to their
corresponding parent in the template - which manifested itself in
strange ways, which is why it took so long to discover.

This migration fixes all parents by assuming there exists a parent page
with the same name on the correct new site. There were a few exceptions
to that assumptions and those were handled manually, se the except
clause.