Rebuild the entire Page MPTT tree after fixing fields
Now that the parent and tree_id fields are fixed, we need to do a
complete rebuild to fix the many inconsistencies caused by the previous
errors.

This WILL cause some random reordering and restructuring of existing
pages, however it will also cause lost pages to re-appear, and hopefully
set the stage for a consistent and correct MPTT tree from hereon out.